changed data type of _runCount to unsigned int
it was previously set as double wich makes no sense for un integer value
thanks to 'bricoleau' on arduino forum for pointing that out

+++ yasm.h
@@ -51,7 +51,7 @@ class YASM{
 		void (*_pLastState)();
 		unsigned long _enterTime;	
 		unsigned long _lastTime;
-		double _runCount;
+		unsigned int _runCount;
 		static void nop(){};
 		bool _isFirstRun;
 };
